getregex/ocr/requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
click==8.1.3
Flask==2.1.2
Flask-RESTful==0.3.9
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
packaging==21.3
Pillow==9.2.0
pycodestyle==2.8.0
pyparsing==3.0.9
pytesseract==0.3.9
pytz==2022.1
six==1.16.0
toml==0.10.2
Werkzeug==2.1.2
zipp==3.8.1
